refactor: separate menu group classes from menu
