+++ src/main/java/kitchenpos/menugroups/tobe/infra/JpaMenuGroupRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.menus.tobe.infra;
+package kitchenpos.menugroups.tobe.infra;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menugroups.tobe.domain.MenuGroup;
+import kitchenpos.menugroups.tobe.domain.MenuGroupRepository;
 
 public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
 }
