+++ src/main/java/kitchenpos/menus/tobe/application/DefaultMenuService.java
@@ -20,8 +20,8 @@ import org.springframework.util.StringUtils;
 import kitchenpos.common.tobe.domain.PurgomalumClient;
 import kitchenpos.menus.tobe.application.dto.MenuCreationRequest;
 import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menugroups.tobe.domain.MenuGroup;
+import kitchenpos.menugroups.tobe.domain.MenuGroupRepository;
 import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.menus.tobe.domain.MenuRepository;
 import kitchenpos.products.tobe.application.ProductService;
