+++ src/test/java/kitchenpos/menus/tobe/application/DefaultMenuServiceTest.java
@@ -25,10 +25,10 @@ import kitchenpos.common.tobe.domain.PurgomalumClient;
 import kitchenpos.common.tobe.infra.FakePurgomalumClient;
 import kitchenpos.menus.tobe.application.dto.MenuCreationRequest;
 import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menugroups.tobe.domain.MenuGroupRepository;
 import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.menus.tobe.domain.MenuRepository;
-import kitchenpos.menus.tobe.infra.InMemoryMenuGroupRepository;
+import kitchenpos.menugroups.tobe.infra.InMemoryMenuGroupRepository;
 import kitchenpos.menus.tobe.infra.InMemoryMenuRepository;
 import kitchenpos.products.tobe.application.DefaultProductService;
 import kitchenpos.products.tobe.application.ProductService;
