+++ src/test/java/kitchenpos/menus/tobe/domain/MenuTest.java
@@ -16,6 +16,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import kitchenpos.menugroups.tobe.domain.MenuGroup;
 import kitchenpos.menus.tobe.application.dto.MenuCreationRequest;
 import kitchenpos.products.tobe.domain.Product;
